requirements.txt
fastapi~=0.88.0
pydantic~=1.10.2
python-dotenv~=0.21.0
PyJWT~=2.6.0
starlette~=0.21.0
bcrypt~=4.0.1
requests~=2.28.1
pymongo~=4.3.3
motor~=3.1.1
httpx~=0.23.1
Jinja2~=3.1.2